backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
